pyproject.toml
[project]
name = "flowstate"
version = "0.1.0"
description = "Ambient stress-aware agent — watches you work, takes over when you hit the wall"
requires-python = ">=3.11"
dependencies = [
    # Agent orchestration
    "railtracks>=1.3.6",
    "claude-agent-sdk>=0.1.50",

    # AI models
    "google-genai>=1.0.0",
    "anthropic>=0.40.0",

    # Webcam + screen capture
    "opencv-python>=4.9.0",
    "mss>=9.0.0",
    "pyautogui>=0.9.54",

    # Browser automation (fallback path)
    "playwright>=1.44.0",

    # Backend server
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sse-starlette>=2.0.0",

    # Auth + rate limiting
    "unkey-py>=0.3.0",

    # Augment Code MCP client
    "mcp>=1.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
]

[project.scripts]
flowstate = "flowstate.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["flowstate"]
